constants changed

+++ src/main/java/io/fourfinanceit/homework/config/Constants.java
@@ -4,6 +4,6 @@ import java.math.BigDecimal;
 
 public class Constants {
 	public static final int MAX_REQUESTS_PER_DAY = 3;
-	public static final BigDecimal MAX_LOAN_AMOUNT = BigDecimal.valueOf(1000);
+	public static final int MAX_LOAN_AMOUNT = 1000;
 	public static final BigDecimal INTEREST_FACTOR_PER_WEEK = BigDecimal.valueOf(1.5);
 }
